refactor : error 패키지 infrastruture 에서 domain 패키지로 이동
