+++ src/main/java/lotto/domain/LottoTicket.java
@@ -1,6 +1,6 @@
 package lotto.domain;
 
-import lotto.infrastructure.error.LottoTicketErrorCode;
+import lotto.domain.error.LottoTicketErrorCode;
 
 import java.util.*;
 
@@ -28,7 +28,7 @@ public class LottoTicket {
     public boolean contains(final int number) {
         return lottoNumbers.contains(number);
     }
-    
+
     private void validateNullOrEmpty(final List<Integer> lottoNumbers) {
         if (Objects.isNull(lottoNumbers) || lottoNumbers.isEmpty()) {
             throw new IllegalArgumentException(LottoTicketErrorCode.NOT_ALLOW_NULL_OR_EMPTY.getMessage());
