+++ src/main/java/lotto/domain/error/LottoRank.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.error;
 
 import java.util.Arrays;
 import java.util.List;
