+++ src/test/java/lotto/domain/LottoRankTest.java
@@ -1,0 +1,5 @@
+package lotto.domain;
+
+class LottoRankTest {
+
+}
