+++ src/test/java/lotto/domain/LottoResultTest.java
@@ -1,7 +1,8 @@
 package lotto.domain;
 
-import lotto.infrastructure.error.LottoCountErrorCode;
-import lotto.infrastructure.error.LottoWinningResultErrorCode;
+import lotto.domain.error.LottoCountErrorCode;
+import lotto.domain.error.LottoRank;
+import lotto.domain.error.LottoWinningResultErrorCode;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
