+++ src/test/java/lotto/domain/WinningLottoTicketTest.java
@@ -1,6 +1,6 @@
 package lotto.domain;
 
-import lotto.infrastructure.error.LottoTicketErrorCode;
+import lotto.domain.error.LottoTicketErrorCode;
 import lotto.infrastructure.util.StringUtils;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
